Modularize session sharing client contract
Separate strict sharing DTO validation from network orchestration while preserving the public client facade.
